(Added `Entity`, `Object` base classes. Added `Player` enum. Added `Ship` class.)
The setup here is similar to the one in the kha-pong repository, but with some slight modifications. An `Object` class now represents something that has a position and size. An `Entity` subclasses `Object` and makes it possible to check for overlaps with points and other entities. In other words, an `Entity` is more of a "physical" object, while an `Object` is something that does not interact with the rest of the game.

The `Ship` is an `Entity`, naturally, as it will have to check for collision against asteroids. A ship will control differently depending on the player (indicated by the `Player` enum). It is not allowed to move backwards outside of the screen (bottom), only forwards (top). When it goes beyond the top of the window it reappars back at the bottom. This behaviour only requires a couple of conditional statements.

Because the original Astro Race arcade game did not use sprites, the ship is drawn with Kha's `graphics2` API. Visually, the ship is constructed from several connected line segments which mimic the original Astro Race spaceship design.

At the project-level, two ships are created, initialised, updated, and drawn.

It's worth noting that all of the game related classes are now located under the `game` package name. This is to prevent possible conflicts with other class names used by Haxe, Kha, or Kha's compilation targets.