(Added `Asteroid`. Added `Group` for easier handling of many asteroid instances. Updated `Object` with `draw()` and `update()` methods.)
The asteroids (debris) in Astro Race are multiple instances of the same class - in this case, an `Asteroid`. The `Group` class makes it easier to work with multiple objects of the same type. It wraps around a typed array and provides methods for updating, drawing, and adding objects to the array representing group elements. `Object` was extended with `update()` and `draw()` methods to facilitate `Group` functionality.

An `Asteroid` is a simple construct. It moves horizontally across the screen at a constant pace, circling back around once it leaves the window bounds. An asteroid is initialised at a random (x, y) position on the screen, with a random direction of horizontal movement.