(Added `Timer`. Updated `Ship` graphic. Updated `Group` and `Asteroid`.)
Astro Race features a game timer in the form of a vertical bar positioned in the middle of the screen. The bar gradually shrinks until it reaches the bottom of the screen, signalising the end of the game. A naive implementation is shown in `Timer` as a line drawn from top to bottom, where the starting y-position moves towards the bottom of the screen with every `update()` call. Ideally the timer would be handled by Kha's scheduler and would feature a callback function which would be executed once the timer finishes.

Collision is also far from ideal, but the small number of entities makes the "brute force" approach a valid one. The asteroid group is iterated through and every asteroid is checked for collision against each of the ships. On collision, the ship position is reinitialised to where they were first positioned.