(Added font asset. Added UI text and score display. Tweaked timer.)
As discussed in the kha-pong repository, Kha requires all of its assets to be loaded by the `Asset` class in some form before they can be used. Because the loading occurs asynchronously, a callback function has to perform the creation of the main game class once all of the assets are loaded and prepared. Then the asset (a font, in this case) can be freely used.